+++ Controllers/Settings.php
@@ -18,6 +18,7 @@ class Settings extends Controller
 {
     private SettingRepository $settingsRepo;
 
+
     /**
      * constructor
      * @access public
@@ -36,8 +37,8 @@ class Settings extends Controller
      */
     public function get(): Response
     {
-        $projectCacheExpiration = (int) ($this->settingsRepo->getSetting('omnisearchsettings.projectscache') ?: 2400);
-        $ticketCacheExpiration = (int) ($this->settingsRepo->getSetting('omnisearchsettings.ticketscache') ?: 1200);
+        $projectCacheExpiration = (int) ($this->settingsRepo->getSetting('omnisearchsettings.projectscache') ?: 60);
+        $ticketCacheExpiration = (int) ($this->settingsRepo->getSetting('omnisearchsettings.ticketscache') ?: 30);
 
         $this->tpl->assign('projectCacheExpiration', $projectCacheExpiration);
         $this->tpl->assign('ticketCacheExpiration', $ticketCacheExpiration);
@@ -52,6 +53,7 @@ class Settings extends Controller
      */
     public function post(array $params): RedirectResponse
     {
+
         $this->settingsRepo->saveSetting('omnisearchsettings.projectscache', (int) ($params['projectCacheExpiration'] ?? 0));
         $this->settingsRepo->saveSetting('omnisearchsettings.ticketscache', (int) ($params['ticketCacheExpiration'] ?? 0));
         $this->tpl->setNotification('The settings were successfully saved.', 'success');